requirements.txt
streamlit
openai
python-docx
faiss-cpu
sentence-transformers
